AGS: Moved arm neon bitmap code to new file.
